refactor getRoomsByFacilityId logic
Previous implementation leveraged a custom RoomRepository method to
query for the rooms within a given facility. New implementation is in
congruence with other implementations of similar logic; now the
RoomService leverages the FacilityService to retrieve the facility
object using its id, this object is used to retrieve its list of rooms.

This implementation is subject to change, following a evaluation of its
performance and to its adherence of the business requirements